requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.0
psycopg[binary]>=3.2.0
psycopg_pool>=3.2.0